requirements.txt
python-telegram-bot[job-queue]
pyVinted
python-dotenv
PyYAML